requirements.txt
Pillow
pytesseract
colorlog